requirements.txt
yt-dlp
gallery-dl
telethon
requests
python-telegram-bot[ext]
